+++ libcontainer/cgroups/devices/ebpf_linux.go
@@ -1,4 +1,4 @@
-package ebpf
+package devices
 
 import (
 	"errors"
