+++ libcontainer/cgroups/devices/v1_test.go
@@ -1,21 +1,34 @@
-package fs
+package devices
 
 import (
+	"os"
+	"path"
 	"testing"
 
+	"github.com/opencontainers/runc/libcontainer/cgroups"
 	"github.com/opencontainers/runc/libcontainer/cgroups/fscommon"
 	"github.com/opencontainers/runc/libcontainer/configs"
 	"github.com/opencontainers/runc/libcontainer/devices"
 )
 
-func TestDevicesSetAllow(t *testing.T) {
-	path := tempDir(t, "devices")
+func init() {
+	testingSkipFinalCheck = true
+	cgroups.TestMode = true
+}
 
-	writeFileContents(t, path, map[string]string{
+func TestSetV1Allow(t *testing.T) {
+	dir := t.TempDir()
+
+	for file, contents := range map[string]string{
 		"devices.allow": "",
 		"devices.deny":  "",
 		"devices.list":  "a *:* rwm",
-	})
+	} {
+		err := os.WriteFile(path.Join(dir, file), []byte(contents), 0o600)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
 
 	r := &configs.Resources{
 		Devices: []*devices.Rule{
@@ -29,13 +42,12 @@ func TestDevicesSetAllow(t *testing.T) {
 		},
 	}
 
-	d := &DevicesGroup{TestingSkipFinalCheck: true}
-	if err := d.Set(path, r); err != nil {
+	if err := setV1(dir, r); err != nil {
 		t.Fatal(err)
 	}
 
 	// The default deny rule must be written.
-	value, err := fscommon.GetCgroupParamString(path, "devices.deny")
+	value, err := fscommon.GetCgroupParamString(dir, "devices.deny")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -44,7 +56,7 @@ func TestDevicesSetAllow(t *testing.T) {
 	}
 
 	// Permitted rule must be written.
-	if value, err := fscommon.GetCgroupParamString(path, "devices.allow"); err != nil {
+	if value, err := fscommon.GetCgroupParamString(dir, "devices.allow"); err != nil {
 		t.Fatal(err)
 	} else if value != "c 1:5 rwm" {
 		t.Errorf("Got the wrong value (%q), set devices.allow failed.", value)
